requirements.txt
python-docx==0.8.11
PyPDF2==3.0.1